Fixed bug: Export disappearing from custom directory
Happened when model_quantity == "1 Format" and export_file_1_ext == export_file_2_ext because it would export the first, then transfer, then attempt to transfer the non-existent export_file_2 but delete the file by the same name in the custom directory first. This way export_file_2 only gets added to the file_list of items to be moved if model_quantity == "2 Formats".