Feature: Use External textures linked to .blend
It may often be the case where a User wants to convert a .blend file that sources textures from many different directories.  This is technically considered an "External" textures source scenario, but it would not convert in the manner expected if all of the linked textures do not exist adjacent to the .blend file.  As such, a User now has the option to use textures already linked to the .blend.  This packs all linked textures into the current Converter.blend and then uses a packed textures workflow in order to then be able to modify the textures while preserving the original, linked textures.