Option to ignore duplicate materials in library
Now an option to avoid marking duplicate materials as assets if the materials already exist as assets in a given library, which is updated between each converted item.  Might be good to extend this functionality to all asset types that can be marked.  Could implement as an additional filter popup when "Ignore Duplicates" property is enabled.