Started expanding asset types to mark, Organized Archive UI
Expanded asset types that can be marked.  Expanded option to ignore duplicate materials into a filter for each asset type.  Added option to "Use Master Collection" for when importing a .blend file with multiple collections, but the User desires to move all other collections into one master collection and only mark the master as an asset (this is default behavior for non-.blend files that are imported).