Rewrote apply_textures_custom, Fixed Bugs: Custom Textures, Mark Assets, Custom Output
Rewrote custom textures algorithm to rely on local copies of textures_temp_dir instead of setting global variables + preserving materials/textures between converted items.  This also fixed a bug that resulted whenever a 2nd or greater model in the batch conversion imported with a material having the same name as a custom material, which bumped the desired custom material with a ".001" suffix and allowed it to be deleted.

Also fixed a bug in mark_assets resulting from not explicitly selecting "(no library)" to give the associated string property the value "NO_LIBRARY".

Finally, fixed a bug resulting from new custom textures algorithm where the master textures_temp_dir was getting moved to the first converted item's directory and not remaining available to be copied to subsequent items' directories.